requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
gradio>=4.0.0
pillow>=10.0.0
pytest>=8.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0